Move the renderView... methods from Router.php to View.php
